parselogs_rpi.py: remove empty and problematic case
Issue: LINCD-18484

This test case now is basically empty. And due to the change in
oe-core's parselogs case, this case is now problematic. Because
there's no common_errors any more. This means when this layer is
in BBLAYERS, any 'bitbake xxx -c testimage' will error out.

(LOCAL REV: NOT UPSTREAM) -- sent to upstream on 20240704

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>